+++ src/Interfaces/Payment/Beneficiary/MerchantBeneficiaryInterface.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace TrueLayer\Interfaces\Beneficiary;
+namespace TrueLayer\Interfaces\Payment\Beneficiary;
 
 use TrueLayer\Interfaces\Remitter\RemitterVerification\RemitterVerificationInterface;
 
 interface MerchantBeneficiaryInterface extends BeneficiaryInterface
 {
     /**
-     * @return string|null
+     * @return string
      */
-    public function getMerchantAccountId(): ?string;
+    public function getMerchantAccountId(): string;
 
     /**
      * @param string $id
@@ -19,6 +19,30 @@ interface MerchantBeneficiaryInterface extends BeneficiaryInterface
      * @return MerchantBeneficiaryInterface
      */
     public function merchantAccountId(string $id): MerchantBeneficiaryInterface;
+
+    /**
+     * @return string|null
+     */
+    public function getAccountHolderName(): ?string;
+
+    /**
+     * @param string $name
+     *
+     * @return $this
+     */
+    public function accountHolderName(string $name): self;
+
+    /**
+     * @return string|null
+     */
+    public function getReference(): ?string;
+
+    /**
+     * @param string $reference
+     *
+     * @return $this
+     */
+    public function reference(string $reference): self;
 
     /**
      * @return RemitterVerificationInterface|null