+++ src/Interfaces/Payout/Beneficiary/BusinessAccountBeneficiaryInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace TrueLayer\Interfaces\Payout\Beneficiary;
+
+interface BusinessAccountBeneficiaryInterface extends BeneficiaryInterface
+{
+}